+++ Source/imgui_ext.h
@@ -47,4 +47,37 @@ namespace HerosInsight::ImGuiExt
         ImGui::PopFont();
         return opened;
     }
+
+    inline bool RadioArray(const char *label, int *current_item, const char *const items[], int items_count)
+    {
+        ImGui::PushFont(Constants::Fonts::button_font);
+        auto size = ImGui::CalcTextSize(label);
+        auto init_cursor = ImGui::GetCursorScreenPos();
+
+        auto new_cursor = init_cursor;
+        new_cursor.x += size.x + ImGui::GetStyle().ItemSpacing.x;
+        ImGui::SetCursorScreenPos(new_cursor);
+        bool pressed_any = false;
+        for (int i = 0; i < items_count; i++)
+        {
+            bool pressed = ImGui::RadioButton(items[i], *current_item == i);
+            if (pressed)
+            {
+                *current_item = i;
+                pressed_any = true;
+            }
+            if (i < items_count - 1)
+                ImGui::SameLine();
+        }
+
+        auto radio_size = ImGui::GetItemRectSize();
+        auto y_offset = (radio_size.y - size.y) / 2;
+        new_cursor = init_cursor;
+        new_cursor.y += y_offset;
+        auto draw_list = ImGui::GetWindowDrawList();
+        draw_list->AddText(new_cursor, ImGui::GetColorU32(ImGuiCol_Text), label);
+
+        ImGui::PopFont();
+        return pressed_any;
+    }
 }