+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 - Fix main menu being too small (the other way this time).
 - Add setting to hide focused character.
+- Add skill ruleset selector (Mixed, PvE, PvP).
 
 ## 0.5.7
 - Fix main menu being too small.